move customisations to tchap folder
