+++ customisations.json
@@ -3,23 +3,23 @@
     "src/components/views/dialogs/CreateRoomDialog.tsx": "src/tchap/components/views/dialogs/TchapCreateRoomDialog.tsx",
     "src/components/views/verification/VerificationComplete.tsx": "src/tchap/components/views/verification/TchapVerificationComplete.tsx",
     "src/components/views/elements/TchapRoomTypeSelector.tsx": "src/tchap/components/views/elements/TchapRoomTypeSelector.tsx",
-    "src/customisations/ComponentVisibility.ts": "src/customisations/TchapComponentVisibility.ts",
-    "src/customisations/Media.ts": "src/customisations/ContentScanningMedia.ts",
-    "src/components/views/messages/DownloadActionButton.tsx": "src/customisations/components/views/messages/ContentScanningDownloadActionButton.tsx",
-    "src/components/views/messages/MAudioBody.tsx": "src/customisations/components/views/messages/ContentScanningAudioBody.tsx",
-    "src/components/views/messages/MFileBody.tsx": "src/customisations/components/views/messages/ContentScanningFileBody.tsx",
-    "src/components/views/messages/MImageBody.tsx": "src/customisations/components/views/messages/ContentScanningImageBody.tsx",
-    "src/components/views/messages/MImageReplyBody.tsx": "src/customisations/components/views/messages/ContentScanningImageReplyBody.tsx",
-    "src/components/views/messages/MStickerBody.tsx": "src/customisations/components/views/messages/ContentScanningStickerBody.tsx",
-    "src/components/views/messages/MVideoBody.tsx": "src/customisations/components/views/messages/ContentScanningVideoBody.tsx",
-    "src/components/views/messages/MVoiceMessageBody.tsx": "src/customisations/components/views/messages/ContentScanningVoiceMessageBody.tsx",
-    "src/components/views/messages/OriginalAudioBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MAudioBody.tsx",
-    "src/components/views/messages/OriginalFileBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MFileBody.tsx",
-    "src/components/views/messages/OriginalImageBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MImageBody.tsx",
-    "src/components/views/messages/OriginalImageReplyBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MImageReplyBody.tsx",
-    "src/components/views/messages/OriginalStickerBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MStickerBody.tsx",
-    "src/components/views/messages/OriginalVideoBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MVideoBody.tsx",
-    "src/components/views/messages/OriginalVoiceMessageBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MVoiceMessageBody.tsx",
+    "src/customisations/ComponentVisibility.ts": "src/tchap/customisations/TchapComponentVisibility.ts",
+    "src/customisations/Media.ts": "src/tchap/customisations/ContentScanningMedia.ts",
+    "src/components/views/messages/DownloadActionButton.tsx": "src/tchap/customisations/components/views/messages/ContentScanningDownloadActionButton.tsx",
+    "src/components/views/messages/MAudioBody.tsx": "src/tchap/customisations/components/views/messages/ContentScanningAudioBody.tsx",
+    "src/components/views/messages/MFileBody.tsx": "src/tchap/customisations/components/views/messages/ContentScanningFileBody.tsx",
+    "src/components/views/messages/MImageBody.tsx": "src/tchap/customisations/components/views/messages/ContentScanningImageBody.tsx",
+    "src/components/views/messages/MImageReplyBody.tsx": "src/tchap/customisations/components/views/messages/ContentScanningImageReplyBody.tsx",
+    "src/components/views/messages/MStickerBody.tsx": "src/tchap/customisations/components/views/messages/ContentScanningStickerBody.tsx",
+    "src/components/views/messages/MVideoBody.tsx": "src/tchap/customisations/components/views/messages/ContentScanningVideoBody.tsx",
+    "src/components/views/messages/MVoiceMessageBody.tsx": "src/tchap/customisations/components/views/messages/ContentScanningVoiceMessageBody.tsx",
+    "src/tchap/components/views/messages/OriginalAudioBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MAudioBody.tsx",
+    "src/tchap/components/views/messages/OriginalFileBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MFileBody.tsx",
+    "src/tchap/components/views/messages/OriginalImageBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MImageBody.tsx",
+    "src/tchap/components/views/messages/OriginalImageReplyBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MImageReplyBody.tsx",
+    "src/tchap/components/views/messages/OriginalStickerBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MStickerBody.tsx",
+    "src/tchap/components/views/messages/OriginalVideoBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MVideoBody.tsx",
+    "src/tchap/components/views/messages/OriginalVoiceMessageBody.tsx": "node_modules/matrix-react-sdk/src/components/views/messages/MVoiceMessageBody.tsx",
     "src/components/views/messages/UnknownBody.tsx": "src/tchap/components/views/messages/TchapUnknownBody.tsx",
     "src/components/views/settings/JoinRuleSettings.tsx": "src/tchap/components/views/settings/TchapJoinRuleSettings.tsx"
 }
