--- src/customisations/README.md
@@ -1,1 +1,0 @@
-../../docs/customisations.md
