+++ src/tchap/customisations/TchapComponentVisibility.ts
